requirements.txt
-i https://pypi.org/simple
alembic==1.13.1;
aniso8601==9.0.1
bcrypt==4.1.2; 
blinker==1.7.0;
click==8.1.7;
faker==23.1.0
flask==3.0.2; 
flask-cors==4.0.0
flask-migrate==4.0.5;
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; 
greenlet==3.0.3;
gunicorn==21.2.0;
itsdangerous==2.1.2;
jinja2==3.1.3; 
mako==1.3.2;
markupsafe==2.1.5;
marshmallow==3.20.2;
packaging==23.2; 
psycopg2-binary==2.9.9;
python-dateutil==2.8.2; 
pytz==2024.1
six==1.16.0; 
sqlalchemy==2.0.26;
sqlalchemy-serializer==1.4.1
typing-extensions==4.9.0; 
werkzeug==3.0.1
Flask-JWT-Extended==4.6.0